move things around and add TODOs
