+++ sql/expression/sort.go
@@ -15,8 +15,6 @@
 package expression
 
 import (
-	"container/heap"
-
 	"github.com/dolthub/go-mysql-server/sql"
 )
 
@@ -42,6 +40,8 @@ func (s *Sorter) Swap(i, j int) {
 func (s *Sorter) IsLesserRow(a, b sql.Row) bool {
 	for _, sf := range s.SortFields {
 		typ := sf.Column.Type(s.Ctx)
+		// TODO: For complex SortFields, like Subqueries, recalculating the value may be costly. We should find some way
+		//  to cache it.
 		av, err := sf.Column.Eval(s.Ctx, a)
 		if err != nil {
 			s.LastError = sql.ErrUnableSort.Wrap(err)
@@ -104,42 +104,4 @@ func (s *ValueRowSorter) Len() int {
 
 func (s *ValueRowSorter) Swap(i, j int) {
 	s.Rows[i], s.Rows[j] = s.Rows[j], s.Rows[i]
-}
-
-// TopRowsHeap implements heap.Interface based on Sorter. It inverts the Less()
-// function so that it can be used to implement TopN. heap.Push() rows into it,
-// and if Len() > MAX; heap.Pop() the current min row. Then, at the end of
-// seeing all the rows, call Rows(). Rows() will return the rows which come
-// back from heap.Pop() in reverse order, correctly restoring the order for the
-// TopN elements.
-type TopRowsHeap struct {
-	Sorter
-}
-
-func (h *TopRowsHeap) Less(i, j int) bool {
-	return !h.Sorter.Less(i, j)
-}
-
-func (h *TopRowsHeap) Push(x interface{}) {
-	h.Sorter.Rows = append(h.Sorter.Rows, x.(sql.Row))
-}
-
-func (h *TopRowsHeap) Pop() interface{} {
-	n := len(h.Sorter.Rows)
-	res := h.Sorter.Rows[n-1]
-	h.Sorter.Rows = h.Sorter.Rows[:n-1]
-	return res
-}
-
-func (h *TopRowsHeap) Rows() ([]sql.Row, error) {
-	if h.LastError != nil {
-		return nil, h.LastError
-	}
-
-	l := h.Len()
-	res := make([]sql.Row, l)
-	for i := l - 1; i >= 0; i-- {
-		res[i] = heap.Pop(h).(sql.Row) // TODO: this is slow
-	}
-	return res, nil
 }
